# server/openapi/skills-registry.v1.yaml
openapi: 3.1.0
info:
  title: CoreAgent Skills Registry API
  version: 1.0.0
  description: |
    User-scoped endpoints require `Authorization: Bearer <JWT>`.
servers:
  - url: http://localhost:4010
paths:
...
  /v1/skills/{skillId}:
    get:
      summary: Get skill details
      parameters:
        - in: path
          name: skillId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Skill details
        "404":
          description: Not found
        "429":
          description: Rate limit exceeded
...
  /v1/admin/skills/publish:
    post:
      summary: Publish or update a skill version
      description: |
        Requires an artifact digest from `/v1/admin/artifacts/upload` (or external `artifactUri`).
        Signature metadata is generated by the registry.
      parameters:
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published
        "401":
          description: Invalid token
        "404":
          description: Admin API disabled
        "503":
          description: Server not configured
  /v1/admin/artifacts/upload:
    post:
      summary: Upload skill artifact to immutable digest storage
      parameters:
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Uploaded
        "400":
          description: Invalid payload
        "401":
          description: Invalid token
        "404":
          description: Admin API disabled
  /v1/admin/skills/{skillId}/versions/{version}/revoke:
    post:
      summary: Revoke a skill version
      parameters:
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
        - in: path
          name: skillId
          required: true
          schema:
            type: string
        - in: path
          name: version
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Revoked
        "401":
          description: Invalid token
        "404":
          description: Not found or admin API disabled
        "503":
          description: Server not configured